# Fase-I/guidelines/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sistema de Análisis de Sentimiento de Reviews – Hotel OLH",
    "description": "API REST para el pipeline de NLP y el dashboard ejecutivo de análisis de sentimiento de reseñas hoteleras. Cubre los servicios: Data Verification & Cleaner, Sentiment Prediction, ABSAService (ABSA sobre tópicos hoteleros → `review_topicos`), Topic Identification (alternativa al ABSA para la misma etapa `topics_identified`), Metric Calculation (`POST /metricas/calcular` y `POST /metricas/recalcular` en MetricProcessor), Dashboard Backend (lecturas `GET /metricas/*`, `POST /metricas/recalcular-semestre` que orquesta recálculos vía MetricProcessor, `GET /metricas/archivos/entrada`, `POST /metricas/archivos/upload` y Reviews Explorer) y el listado filtrado de reviews.",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo Tesis OLH",
      "email": "admin@olh.com"
    }
  },
  "servers": [
    {
      "url": "https://api.olh.com",
      "description": "Producción"
    },
    {
      "url": "http://localhost:8000",
      "description": "Desarrollo local (genérico)"
    },
    {
      "url": "http://localhost:5002",
      "description": "Dashboard Backend (local): GET /api/v1/metricas/*, POST /api/v1/metricas/recalcular-semestre, GET /api/v1/metricas/archivos/entrada y POST /api/v1/metricas/archivos/upload"
    },
    {
      "url": "http://localhost:5004",
...
    },
    {
      "url": "http://localhost:5003",
...
    },
    {
...
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/metricas/recalcular-semestre": {
      "post": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Recalcular métricas del último semestre (Dashboard Backend)",
        "description": "Recibe `hotel_id` y orquesta seis llamadas secuenciales en orden cronológico ascendente a MetricProcessor `POST /api/v1/metricas/recalcular` para los últimos 6 meses naturales incluyendo el mes actual (rolling). Devuelve un resumen (`exitos`, `fallos`, `fallidos`). Requiere JWT. Configuración: `METRIC_PROCESSOR_URL`, `METRIC_PROCESSOR_TIMEOUT_SECONDS` en el backend del dashboard.",
        "operationId": "postMetricasRecalcularSemestre",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecalcularSemestreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Semestre procesado (puede incluir fallos parciales por mes sin fila en `metricas_globales_mensual`)",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RecalcularSemestreResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Body inválido (`hotel_id`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                },
                "example": {
                  "ok": false,
                  "data": null,
                  "error": {
                    "codigo": "BODY_INVALIDO",
                    "mensaje": "hotel_id debe ser un entero mayor a 0"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "description": "MetricProcessor no alcanzable o timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                },
                "example": {
                  "ok": false,
                  "data": null,
                  "error": {
                    "codigo": "METRIC_PROCESSOR_INDISPONIBLE",
                    "mensaje": "No se pudo conectar con MetricProcessor."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/metricas/kpis": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "KPIs mensuales agregados (Dashboard Backend)",
        "description": "Devuelve KPIs mensuales leyendo `metricas_globales_mensual`.",
        "operationId": "getMetricasKpisMensual",
        "parameters": [
          {
            "$ref": "#/components/parameters/HotelId"
          },
          {
            "$ref": "#/components/parameters/Anio"
          },
          {
            "$ref": "#/components/parameters/Mes"
          }
        ],
        "responses": {
          "200": {
            "description": "KPIs del período",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/KpisResponse"
                        }
                      }
                    }
                  ]
                },
                "examples": {
                  "conDatos": {
                    "summary": "Período con métricas calculadas",
                    "value": {
                      "ok": true,
                      "data": {
                        "sentimiento_promedio": {
                          "score": 68.4,
                          "cambio_pct": 2.1,
                          "tendencia": "up"
                        },
                        "reviews_analizadas": 1432,
                        "topicos_con_alerta": 2
                      },
                      "error": null
                    }
                  },
                  "sinDatos": {
                    "summary": "Período sin métricas",
                    "value": {
                      "ok": true,
                      "data": {
                        "sentimiento_promedio": {
                          "score": 0,
                          "cambio_pct": 0,
                          "tendencia": "stable"
                        },
                        "reviews_analizadas": 0,
                        "topicos_con_alerta": 0
                      },
                      "error": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos (hotel_id, anio, mes)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                },
                "example": {
                  "ok": false,
                  "data": null,
                  "error": {
                    "codigo": "PARAMETROS_INVALIDOS",
                    "mensaje": "'anio' debe ser un entero mayor o igual a 2020"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/metricas/sentimiento-mensual": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Serie mensual de sentimiento (Dashboard Backend)",
        "description": "Devuelve la evolución mensual del score de sentimiento leyendo `metricas_globales_mensual`.",
        "operationId": "getMetricasSentimientoMensual",
        "parameters": [
          {
            "$ref": "#/components/parameters/HotelId"
          },
          {
            "name": "meses",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 12,
              "minimum": 1,
              "maximum": 24
            },
            "description": "Últimos N meses incluyendo el actual. Se ignora si se usan `desde` y `hasta`."
          },
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])$",
              "example": "2025-01"
            },
            "description": "Inicio del rango en formato YYYY-MM. Requiere `hasta`."
          },
          {
            "name": "hasta",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])$",
              "example": "2025-11"
            },
            "description": "Fin del rango en formato YYYY-MM. Requiere `desde`."
          }
        ],
        "responses": {
          "200": {
            "description": "Serie mensual de sentimiento",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PuntoSentimientoMensual"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos (hotel_id, meses o rango desde/hasta)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                },
                "example": {
                  "ok": false,
                  "data": null,
                  "error": {
                    "codigo": "PARAMETROS_INVALIDOS",
                    "mensaje": "'hotel_id' es requerido"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/metricas/topicos": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Métricas por tópico del mes (Dashboard Backend)",
        "description": "Devuelve métricas por tópico leyendo `metricas_topico_mensual` y `topicos`.",
        "operationId": "getMetricasTopicosMensual",
        "parameters": [
          {
            "$ref": "#/components/parameters/HotelId"
          },
          {
            "$ref": "#/components/parameters/Anio"
          },
          {
            "$ref": "#/components/parameters/Mes"
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "clave",
                "adicional"
              ]
            },
            "description": "Filtrar por tipo de tópico. Vacío = todos."
          },
          {
            "name": "solo_alertas",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Si es true, retorna solo tópicos con alerta activa"
          }
        ],
        "responses": {
          "200": {
            "description": "Métricas de tópicos del período",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TopicoMetrica"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos (hotel_id, anio, mes, tipo o solo_alertas)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                },
                "example": {
                  "ok": false,
                  "data": null,
                  "error": {
                    "codigo": "PARAMETROS_INVALIDOS",
                    "mensaje": "El parámetro 'mes' es requerido"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/metricas/topicos/top5": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Top 5 tópicos más críticos (Dashboard Backend)",
        "description": "Devuelve los 5 tópicos con menor score del período leyendo `metricas_topico_mensual` y `topicos`.",
        "operationId": "getMetricasTopicosTop5Mensual",
        "parameters": [
          {
            "$ref": "#/components/parameters/HotelId"
          },
          {
            "$ref": "#/components/parameters/Anio"
          },
          {
            "$ref": "#/components/parameters/Mes"
          }
        ],
        "responses": {
          "200": {
            "description": "Top 5 tópicos más críticos del período",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "maxItems": 5,
                          "items": {
                            "$ref": "#/components/schemas/TopicoTop5"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Parámetros inválidos (hotel_id, anio, mes)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeError"
                },
                "example": {
                  "ok": false,
                  "data": null,
                  "error": {
                    "codigo": "PARAMETROS_INVALIDOS",
                    "mensaje": "El parámetro 'mes' debe estar entre 1 y 12"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/v1/metricas/topicos/{slug}/detalle": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Detalle de un tópico para el modal",
        "description": "Detalle completo de un tópico: score, tendencia, desglose de menciones y fragmentos reales de reviews. Alimenta el modal de detalle del dashboard.",
        "operationId": "getTopicoDetalle",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "ruido"
            },
            "description": "Identificador único del tópico (ej: ruido, limpieza, wifi)"
          },
          {
            "$ref": "#/components/parameters/HotelId"
          },
          {
            "$ref": "#/components/parameters/Anio"
          },
          {
            "$ref": "#/components/parameters/Mes"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del tópico con fragmentos de reviews",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TopicoDetalle"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "HotelId": {
        "name": "hotel_id",
        "in": "query",
        "required": true,
        "schema": {
          "type": "integer"
        },
        "description": "ID del hotel"
      },
...
    },
...
    "schemas": {
...
      "RecalcularSemestreRequest": {
        "type": "object",
        "required": [
          "hotel_id"
        ],
        "properties": {
          "hotel_id": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "example": 1
          }
        }
      },
...
    }
  }
}